requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.26.0
apscheduler==3.8.1
nest_asyncio==1.5.4
pytz==2023.3
pytest==7.1.2
watchdog[watchmedo]==3.0.0